requirements-dev: Add requirements.txt constraint
Add "-c requirements.txt" line to requirements-dev.in to make sure that
requirements-dev.txt won't contain any conflicting versions of the
packages in requirements.txt.

For example, consider situation that a package in requirements-dev.in
requires a package P which has versions 1.1 and 1.2.  It's possible that
version 1.1 is already pinned in requirements.txt, but without this
constraint line the newer version 1.2 would be pinned in
requirements-dev.txt.