Add lease ID helper functions
These are needed to translate lease ID strings, so that for instance
A1122-4 becomes A1, 1, 22 and 4 (this is done by get_identifier_parts).

The legacy database uses the first and second parts, in the example
A1122 and 4, to determine relations between tables, so
get_from_identifier can be used to get the corresponding lease in the
new system from a reference in the old one.