Add virre integration
Virre is an API by Finnish Patent and Registration Office.

We are only using some of the existing endpoints. If the endpoint
generates a PDF file, we fetch it from the response and pass it
to the UI. If the endpoint returns xml data, we pass its parsed
json version as is to the frontend (proxy approach).

Other endpoints can be added later, following the existing
pattern.