Setup Mypy type checking
Add Mypy for statical type checking the batchrun app.

The included run-type-checks script can be used to run the type checks.

This includes so called stub files for making some of the used libraries
type check compatible.  Most of the stubs are bundled with mypy, but we
also install django-stubs and djangorestframework-stubs as separate
packages.  In addition some additional stub files are added to stubs
directory for those libraries that don't yet have stubs available in
Mypy or PyPI.