Overhaul rent share calculation and invoice generation
Previously the rent for a given period was calculated as a one amount
and the amount was then shared between tenants by the tenants
share.

Later it was decided that there should be possibility to share
rents for the different intended uses by share that is not the
same as the tenant share.

In the new system the rent is calculated for every intended use
separately and then the separate amounts are shared between the
tenants.