Improve the plan units modification information
The plan units modified_at field cannot change every time when the save is called. It needs to look at the fields first, that are there any changes and only after that the modified_at field can be updated. This makes it possible to compare the similar plan units by the modified_at field.