Fix async report execution
The async tasks of reports fails to the pickle error silenty and the requester doesn't get the report via email.

The causes was in the Django forms. The renderer isn't supported for the pickle so it crashes to this error:
PicklingError at /v1/report/rent_forecast/
Can't pickle <function paginator_number at 0x7f793b308400>: it's not the same object as django.contrib.admin.templatetags.admin_list.paginator_number

The renderer of form has been turned to None as the report feature doesn't require any rendering. This solves the pickle problem in async tasks in reports.