settings: Set better default value for DATABASE_URL
If PostgreSQL is running on the localhost the best way to connect to it
is via the UNIX socket, which doesn't even need username and password,
since it can utilize the peer authentication.  Make this as the default
value of the DATABASE_URL.

This makes it easier to develop MVJ on local machine without Docker.