Modify LandUseAgreementType model
Land use agreement type has nothing to do with the
LeaseType model which has been used as the base for
this model. Remove the obsolete fields from the
LandUseAgreementModel and make the identifier non-unique
since it will usually be 'MA'.

Also update the django fixtures for the land use agreement
types and update some model field translations and comments.