batchrun: Fix log message decoding
Sometimes log messages are split between the bytes which make the UTF-8
encoding of a single character and that causes decoding error.

E.g. "ö" is encoded as two bytes C3 B6 in UTF-8 and if a log message is
split between these bytes to two separate log messages, then the first
message will end to byte C3 and the next message starts with byte B6.
Both of these messages will then get corrupted, because those
undecodable bytes are coded with the replace character U+FFFD.

Make the decoding handle this case by using a stateful decoder which
stores the extra byte from the first message to its internal state and
then correctly retrieves it from the state when decoding the second
message.  This means that the "ö" split like that would be decoded as
the first character of the second message.