batchrun: Add JobRunLog model
Add a new model "JobRunLog", which can store a compact version of all of
the JobRunLogEntry objects of a JobRun.  All of the log entries will be
concatenated to a single text field.  This makes it compressable for
PostgreSQL and therefore consumes a lot less space than the separate
entries.

Tested that about 28 GiB of log entries can be stored to just 1 GiB as
JobRunLog objects.