batchrun: Improve memory usage of JobRunLogAdmin view
When generating the queryset for the "Logs" admin view Django fetches
also the big blob fields "content" and "entry_data" of the JobRunLog
objects by default, even though those fields are not even visible in the
list view.  That makes the view rendering memory requirements quite
high.  On the QA server it fails to render the view and instead raises a
MemoryError.

Fix this issue by deferring the "content" and "entry_data" fields so
that they won't be fetched from the database for the list view.