requirements: Update drf-oidc-auth to 0.10.0
Update the drf-oidc-auth library to version 0.10.0, because there's a
bug in 0.9.0 version that would make the production environment totally
broken for users logged in via Tunnistamo.

The version 0.10.0 is already installed to production to solve the
issue.  It's good to have the requirements.txt match to the version
installed so that it won't be downgraded again on next update.