batchrun: Convert entry_data to text json field
Django uses the "jsonb" data type for the JSON fields by default, but
that does not work for the log table, since for very long logs the
metadata could be quite big and there is a 256 MiB limit for the jsonb
data type in PostgreSQL.

Also the disk usage of the table also dropped by 6% when the type was
changed from "jsonb" to "json".  Downside is slower JSON operations on
queries, but those are not so much needed for the log storage, since the
most important metadata is already stored as table columns (entry_count,
error_count, start and end times).