batchrun: Add PostgreSQL procedure for log compacting
Add a migration file that creates a plpgsql stored procedure for
generating compacted logs from the log entries.

See documentation of the procedure from the SQL file.

The migration file reads the actual code from another file so that it is
possible to edit it as SQL file more easily -- e.g. syntax highlighting
works much better when the SQL code is in a .sql file rather than
embedded into Python code.

NOTE: Since the code is written with plpgsql and is using a lot of
PostgreSQL specific functions, this change will effectively make the
batchrun app to no longer be database agnostic, but rather requiring the
DBMS to be PostgreSQL.  If support for other DBMSs is needed, this part
would have to be revisited.