batchrun: Use the plpgsql procedure to compact logs
Update the log compacting Python code to utilize the stored procedure to
do the main work.  The log table row is precreated in the Python code
and then its content will be filled by calling the procedure.

Main reason for this change was optimizing the memory footprint of the
Python code and/or the memory pressure pushed to the PostgreSQL process.
Before this change, the compacting failed for some of the job runs which
had 4.5 million log entries.  The Python code consumed more than 4 GiB
of memory and together with the PostgreSQL the total memory usage was
eating all RAM and swap causing one of them to be OOM killed before
completion.

This solution takes away the memory pressure from the Python code, since
the log entries are no longer transmitted to the Python process at all
when compating.  All heavy work is done in the PostgreSQL process by the
stored procedure.  The stored procedure however uses temporary tables to
store the intermediate results, which allows it to work with less
memory, since the temporary table can utilize the normal database disk
space, if needed.

I tested the stored procedure with a test set of 10 million log entries
and limiting the PostgreSQL process memory to 3 GiB and it was still
able to process it.  The processing time was also reduced for that set
from 240 s to 75 s, i.e. almost 70% decrease.  (It needed about 8 GiB of
memory to process that 10M entries with the old code.)