Revise recursive_get_related_skip_relations for models
The feature relates to auditlog endpoint that shows all the related changes to an instance e.g. areasearch, lease, decision.
Setting the on models values mostly for reverse relations so that the check doesn't end up in loops and hoops, disconnecting user model just in case. We don't want the recursion to start collecting everything related to a user.
Forms and Plotsearch was missing these completely, this is my best guess of how it should be
Not all of these models might be in auditlog, but this helps in case they are getting added