chore: update django and djangorestframework along with other django dependencies
djangorestframework to 3.14.0, due to 3.15.0 having issues with enumfields.Enum (django-enumfields2), it is unable to serialize to JSON. Possibly requires a fix in that package, or the original fork.
Add REST_FRAMEWORK["DEFAULT_VERSION"] in settings, some breaking change that was not documented.
django-safedelete to 1.3.3 generated migrations for apps using it.