Fix bug 'ExcelRow' object has no attribute 'keys'
This while loop is expected to identify each ExcelRow, and skip those in
favour of dict-format rows. Because `data` is a list, `lookup_row_num in
data` incorrectly checked whether the integer value of lookup row number
is a value on that row. This shorted the if-statement, which on a later
line allowed invocation of method ExcelRow.keys(), which doesn't exist.