Add management function to validate sanitizerconfig
This management command checks the file `.sanitizerconfig`, and lists
all models and their fields that are present in current Django models,
but missing from the configuration.

Run it before taking a sanitized dump from a database, so you know
what to update in the configuration.

The existing validation command `check_sanitizerconfig` from
python-sanitized-dump unnecessarily complained about models with
skip_rows strategy, and other miscellaneous issues, even if the
configuration was created with `init_sanitizer`.