remove BankHoliday model and associated code
The model was not updated since 2021 and thus was not working as intended.
According to the information available SAP handles bank holidays when generating invoices, and therefore it is not critical for this system to handle it. And it has been working anyway without any updates to the BankHoliday model.