Revert "Sort invoice rows depending on their type"
This reverts the following commits:

* 57420083ed0e1c3d7baff9353226160d31b65d25
* ae102fbc30085d6dae44e5ada3985e0d11fbff2f
* c15b5e450581d6eb67dc4576a793dcb03a90200c

This method was unreliable for determining invoice row order in invoice
exports. Automatic invoice generation is complicated enough to not be
able to deduce the row's intended type from position in code alone. The
invoice row ordering will instead be deduced from row's amount at the
place of sorting when generating invoice export XMLs.